requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.109.2
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.25
starlette==0.36.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.27.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
